SphericalMAE/requirements.txt

timm==0.4.12
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
# deepspeed==0.4.0
scipy
nilearn
pyvista
